Move BlobResourceHandleBase class to its own file https://bugs.webkit.org/show_bug.cgi?id=301156
Reviewed by Darin Adler.

Move BlobResourceHandleBase class to its own file for clarity. No behavior change.

* Source/WebCore/Headers.cmake:
* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/platform/network/BlobResourceHandle.cpp:
(WebCore::BlobResourceHandleBase::BlobResourceHandleBase): Deleted.
(WebCore::BlobResourceHandleBase::adjustAndValidateRangeBounds): Deleted.
(WebCore::BlobResourceHandleBase::closeFileIfOpen): Deleted.
(WebCore::BlobResourceHandleBase::clearAsyncStream): Deleted.
(WebCore::BlobResourceHandleBase::blobData const): Deleted.
(WebCore::BlobResourceHandleBase::syncStream const): Deleted.
(WebCore::BlobResourceHandleBase::asyncStream const): Deleted.
(WebCore::BlobResourceHandleBase::start): Deleted.
(WebCore::BlobResourceHandleBase::doStart): Deleted.
(WebCore::BlobResourceHandleBase::getSizeForNext): Deleted.
(WebCore::BlobResourceHandleBase::didGetSize): Deleted.
(WebCore::BlobResourceHandleBase::seek): Deleted.
(WebCore::BlobResourceHandleBase::readAsync): Deleted.
(WebCore::BlobResourceHandleBase::readDataAsync): Deleted.
(WebCore::BlobResourceHandleBase::readFileAsync): Deleted.
(WebCore::BlobResourceHandleBase::didOpen): Deleted.
(WebCore::BlobResourceHandleBase::didRead): Deleted.
(WebCore::BlobResourceHandleBase::consumeData): Deleted.
(WebCore::BlobResourceHandleBase::dispatchDidReceiveResponse): Deleted.
* Source/WebCore/platform/network/BlobResourceHandle.h:
(WebCore::BlobResourceHandleBase::isFileOpen const): Deleted.
(WebCore::BlobResourceHandleBase::setIsFileOpen): Deleted.
(WebCore::BlobResourceHandleBase::async const): Deleted.
(WebCore::BlobResourceHandleBase::totalSize): Deleted.
(WebCore::BlobResourceHandleBase::totalRemainingSize const): Deleted.
(WebCore::BlobResourceHandleBase::currentItemReadSize const): Deleted.
(WebCore::BlobResourceHandleBase::setCurrentItemReadSize): Deleted.
(WebCore::BlobResourceHandleBase::decrementTotalRemainingSizeBy): Deleted.
(WebCore::BlobResourceHandleBase::readItemCount const): Deleted.
(WebCore::BlobResourceHandleBase::incrementReadItemCount): Deleted.
(WebCore::BlobResourceHandleBase::lengthOfItemBeingRead const): Deleted.
(WebCore::BlobResourceHandleBase::buffer): Deleted.
(WebCore::BlobResourceHandleBase::buffer const): Deleted.
(WebCore::BlobResourceHandleBase::shouldAbortDispatchDidReceiveResponse): Deleted.
(WebCore::BlobResourceHandleBase::clearStream): Deleted.
* Source/WebCore/platform/network/BlobResourceHandleBase.cpp: Copied from Source/WebCore/platform/network/BlobResourceHandle.cpp.
(WebCore::BlobResourceHandleBase::BlobResourceHandleBase):
(WebCore::BlobResourceHandleBase::adjustAndValidateRangeBounds):
(WebCore::BlobResourceHandleBase::closeFileIfOpen):
(WebCore::BlobResourceHandleBase::clearAsyncStream):
(WebCore::BlobResourceHandleBase::blobData const):
(WebCore::BlobResourceHandleBase::syncStream const):
(WebCore::BlobResourceHandleBase::asyncStream const):
(WebCore::BlobResourceHandleBase::start):
(WebCore::BlobResourceHandleBase::doStart):
(WebCore::BlobResourceHandleBase::getSizeForNext):
(WebCore::BlobResourceHandleBase::didGetSize):
(WebCore::BlobResourceHandleBase::seek):
(WebCore::BlobResourceHandleBase::readAsync):
(WebCore::BlobResourceHandleBase::readDataAsync):
(WebCore::BlobResourceHandleBase::readFileAsync):
(WebCore::BlobResourceHandleBase::didOpen):
(WebCore::BlobResourceHandleBase::didRead):
(WebCore::BlobResourceHandleBase::consumeData):
(WebCore::BlobResourceHandleBase::dispatchDidReceiveResponse):
* Source/WebCore/platform/network/BlobResourceHandleBase.h: Copied from Source/WebCore/platform/network/BlobResourceHandle.h.
(WebCore::BlobResourceHandleBase::isFileOpen const):
(WebCore::BlobResourceHandleBase::setIsFileOpen):
(WebCore::BlobResourceHandleBase::async const):
(WebCore::BlobResourceHandleBase::totalSize):
(WebCore::BlobResourceHandleBase::totalRemainingSize const):
(WebCore::BlobResourceHandleBase::currentItemReadSize const):
(WebCore::BlobResourceHandleBase::setCurrentItemReadSize):
(WebCore::BlobResourceHandleBase::decrementTotalRemainingSizeBy):
(WebCore::BlobResourceHandleBase::readItemCount const):
(WebCore::BlobResourceHandleBase::incrementReadItemCount):
(WebCore::BlobResourceHandleBase::lengthOfItemBeingRead const):
(WebCore::BlobResourceHandleBase::buffer):
(WebCore::BlobResourceHandleBase::buffer const):
(WebCore::BlobResourceHandleBase::shouldAbortDispatchDidReceiveResponse):
(WebCore::BlobResourceHandleBase::clearStream):
* Source/WebKit/NetworkProcess/NetworkDataTaskBlob.h:

Canonical link: https://commits.webkit.org/301868@main
